# Proof of concept - Do not merge - Tests to decide labels field model